requirements.txt
RealtimeSTT
crewai
streamlit
sounddevice
openai-whisper